Big in-game U.I update (read description).
This is a big commit update for the in-game User Interface. I have, over the past few days, created the improved UI art (coloured/finished) for the other half of the in-game UI - above that I have implemented half of that (both the 'auto kinetic gun' and 'resource well') into the Unity file (all that needs to finish both them features off now is the coding).